+++ api/src/entity/users.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "users")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub user_id: i32,
+    pub user_id: i64,
     #[sea_orm(unique)]
     pub username: String,
     pub password: String,
